+++ Benchmarking/PerfCallMethodUpdate2.cs
@@ -65,7 +65,7 @@ namespace Benchmarking
             //SETUP
             using (var context = new EfCoreContext(_options))
             {
-                var service = new CrudServices<EfCoreContext>(context, _utData.Wrapped);
+                var service = new CrudServices<EfCoreContext>(context, _utData.ConfigAndMapper);
 
                 //ATTEMPT
                 var newDate = new DateTime(2000, 1, 1).AddDays(_incdDay++);
@@ -103,7 +103,7 @@ namespace Benchmarking
             //SETUP
             using (var context = new EfCoreContext(_options))
             {
-                var service = new CrudServices<EfCoreContext>(context, _utData.Wrapped);
+                var service = new CrudServices<EfCoreContext>(context, _utData.ConfigAndMapper);
 
                 //ATTEMPT
                 var newDate = new DateTime(2000, 1, 1).AddDays(_incdDay++);