+++ GenericServices/Configuration/IGenericServicesConfig.cs
@@ -2,6 +2,8 @@
 // Licensed under MIT licence. See License.txt in the project root for license information.
 
 using System;
+using System.ComponentModel.DataAnnotations;
+using Microsoft.EntityFrameworkCore;
 
 namespace GenericServices.Configuration
 {
@@ -15,5 +17,14 @@ namespace GenericServices.Configuration
         /// The DefaultNameMatcher only handles names that are exactly the same, apart from the given name can be camelCase
         /// </summary>
         MatchNameAndType NameMatcher { get; }
+
+        /// <summary>
+        /// When SaveChangesWithValidation is called if there is a DbUpdateException then this method
+        /// is called. If it returns null then the error is rethrown, but if it returns a ValidationResult
+        /// then that is turned into a error message that is shown to the user via the IBizActionStatus
+        /// See section 10.7.3 of my book "Entity Framework Core in Action" on how to use this to turn
+        /// SQL errors into user-friendly errors
+        /// </summary>
+        Func<DbUpdateException, ValidationResult> SqlErrorHandler { get; }
     }
 }