+++ Tests/UnitTests/GenericServicesPublic/TestNestedDtos.cs
@@ -32,7 +32,7 @@ namespace Tests.UnitTests.GenericServicesPublic
 
                 var utData = context.SetupSingleDtoAndEntities<BookListNestedV1Dto>();
                 utData.AddSingleDto<AuthorNestedV1Dto>();
-                var service = new CrudServices(context, utData.Wrapped);
+                var service = new CrudServices(context, utData.ConfigAndMapper);
 
                 //ATTEMPT
                 var dto = service.ReadSingle<BookListNestedV1Dto>(1);
@@ -59,7 +59,7 @@ namespace Tests.UnitTests.GenericServicesPublic
 
                 var utData = context.SetupSingleDtoAndEntities<BookListNestedV2Dto>();
                 utData.AddSingleDto<AuthorNestedV2Dto>();
-                var service = new CrudServices(context, utData.Wrapped);
+                var service = new CrudServices(context, utData.ConfigAndMapper);
 
                 //ATTEMPT
                 var dto = service.ReadSingle<BookListNestedV2Dto>(1);