+++ Tests/UnitTests/GenericServicesPublicAsync/TestReadViaDtoWithPreQuery.cs
@@ -34,7 +34,7 @@ namespace Tests.UnitTests.GenericServicesPublicAsync
                 context.SeedDatabaseFourBooks();
 
                 var utData = context.SetupSingleDtoAndEntities<BookTitle>();
-                var service = new CrudServices(context, utData.Wrapped);
+                var service = new CrudServices(context, utData.ConfigAndMapper);
 
                 //ATTEMPT
                 var list = await service.ReadManyWithPreQueryNoTracked<Book, BookTitle>(books =>